Create folders in ...dto for better structure
