+++ src/main/java/com/school/controller/HomeController.java
@@ -1,10 +1,6 @@
 package com.school.controller;
 
 import com.school.dto.HomeResponse;
-import com.school.dto.SubjectResponse;
-import com.school.dto.TeacherResponseToGet;
-import com.school.models.Subject;
-import com.school.models.Teacher;
 import com.school.service.StudentService;
 import com.school.service.SubjectService;
 import com.school.service.TeacherService;
@@ -14,9 +10,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @RestController
 @RequestMapping("/home")
