+++ src/main/java/com/school/dto/student/StudentRequest.java
@@ -1,4 +1,4 @@
-package com.school.dto;
+package com.school.dto.student;
 
 import com.school.models.Student;
 import com.school.models.User;
