+++ src/main/java/com/school/dto/subject/SubjectRequest.java
@@ -1,4 +1,4 @@
-package com.school.dto;
+package com.school.dto.subject;
 
 import lombok.Data;
