+++ src/main/java/com/school/dto/teacher/TeacherResponseAll.java
@@ -1,23 +1,24 @@
-package com.school.dto;
+package com.school.dto.teacher;
 
 import com.school.models.Teacher;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
-@Getter
-@Setter
 @Data
-public class TeacherResponseToGet {
+public class TeacherResponseAll {
+    private long id;
     private String firstName;
     private String lastName;
     private String email;
-    private long id;
+    private String[] subjects;
 
-    public TeacherResponseToGet(Teacher teacher){
+    public TeacherResponseAll(Teacher teacher){
         this.firstName = teacher.getUser().getFirstName();
         this.lastName = teacher.getUser().getLastName();
         this.email = teacher.getUser().getEmail();
         this.id = teacher.getId();
+        this.subjects = new String[teacher.getSubjects().size()];
+        for(int i = 0; i < subjects.length; i++){
+            subjects[i] = teacher.getSubjects().get(i).getName();
+        }
     }
 }
