+++ src/main/java/com/school/dto/teacher/TeacherResponseToGet.java
@@ -1,24 +1,23 @@
-package com.school.dto;
+package com.school.dto.teacher;
 
 import com.school.models.Teacher;
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
+@Getter
+@Setter
 @Data
-public class TeacherResponseAll {
-    private long id;
+public class TeacherResponseToGet {
     private String firstName;
     private String lastName;
     private String email;
-    private String[] subjects;
+    private long id;
 
-    public TeacherResponseAll(Teacher teacher){
+    public TeacherResponseToGet(Teacher teacher){
         this.firstName = teacher.getUser().getFirstName();
         this.lastName = teacher.getUser().getLastName();
         this.email = teacher.getUser().getEmail();
         this.id = teacher.getId();
-        this.subjects = new String[teacher.getSubjects().size()];
-        for(int i = 0; i < subjects.length; i++){
-            subjects[i] = teacher.getSubjects().get(i).getName();
-        }
     }
 }
